consider journal/manager mismatches failures when fuzzing (#20854)
When fuzzing, we currently always consider SIEs and timeouts a failure.
We want to use fuzzing to catch journaling errors as well, so make those
always a failure too in addition to those other categories.

I've been using this locally for a while, and probably makes sense to
use this for all fuzz testing.