Wrap some transform/source_eval errors (#20871)
We suspect that propagating errors directly from
`CallbacksClient.Invoke` to the return value of
`ResourceMonitorServer.RegisterResource` is obscuring where grpc issues
are actually happening. For example if `Invoke` returns a grpc error
`NotImplemented` and we directly return that from `RegisterResource` the
resource monitor client is going to think `RegisterResource` itself is
not implemented. This changes the code so we have at least one layer of
wrapping around the grpc error before returning it from another grpc
method.

While at it I also wrapped a few other unwrapped errors to make it more
clear what failed if any of them happen.

---------

Co-authored-by: Julien Poissonnier <julien@caffeine.lu>