return a new state from a refresh step when resource is pending replacement (#20812)
Since https://github.com/pulumi/pulumi/pull/20781, we return `nil` in
the completion source as the new state, for a resource that is
refreshed and has been pending replacement.

However this is not quite right, as subsequent steps then don't know
about the existance of that resource anymore, and thus ignore it. This
doesn't work when there's a failed replacement, leaving a resource in
such a state, and the user subsequently uses `up --refresh
--run-program`.

Because the subsequent replacement step after the refresh doesn't know
about the resource pending replacement anymore, it doesn't take it
into account, and creates a new resource, despite the old one not
being deleted. Instead, since we want the resource to stay the same in
the refresh step, we copy the resource, and use that one.

Note that we can't just re-use `s.new` here, because that can be the
same pointer as `s.old`, which is not allowed by the system. Copying
feels like the right thing to do, as we "refresh" the resource this
way, and it works with the rest of the system.

Fixes https://github.com/pulumi/pulumi/issues/20251