refreshv2: only rebuild base state if we have a refresh  (#20861)
In refreshV2, depending on how the program changes, and what resources
are targeted, we might not end up actually refreshing any resources,
but rather just end up with a bunch of same steps.

In particular, this can happen if a resource that's marked as `Delete`
is targeted.

If that happens, we end up incorrectly removing a parent from the
state, because the resource is marked as `Delete`.  The regular
snapshot manager and the journaler handle this correctly, as they only
undangle the parent if there is a refresh step, but we don't handle
this correctly in `rebuildBaseState`.  Don't run `rebuildBaseState` if
there are no refreshes, since it should not be necessary.

This makes sure we get the same result from all the snapshot
implementations in these cases.

This feels like a rather contrived example, but the fuzzer found it, so
it's better to fix it. Note that this does not have any effect on the
snapshot generated from the journaler or the normal snapshot manager,
only the test journal is affected.

Fixes https://github.com/pulumi/pulumi/issues/20862