mark tests using dep_tracker as asyncio (#20876)
Tests using the `dep_tracker` need an event loop to be running to work
correctly. This used to work when we didn't run the tests in parallel,
because we already set up the event loop somehow, but now that we're
running tests in parallel, we need to mark them explicitly for pytest to
start an event loop. Do so to hopefully avoid the flakyness we've been
seeing in CI recently.

Fixes https://github.com/pulumi/pulumi/issues/20869 hopefully