Log gRPC requests before invoking them in gRPC debug interceptor (#20817)
When using `PULUMI_DEBUG_GRPC`, the log entries only record full
operations after they are completed with their requests and responses.
This is not ideal if an operation is blocked/stuck, it won't show up in
the logs.

This PR makes it such that we log two entries per operation
- entry with request and + field `progress: request_started` before
invoking the operation
- entry with request and response + field `progress: response_completed`
after invoking the operation