Use immutable results in simplified invokes (#20877)
This was never end to end tested when first merged, but the dotnet sdk
has limitations on what types are allowed in the return type. So this
changes "scalar invokes" to use the acceptable "Immutable" variants in
codegen, which we can then pull into the dotnet repo to test with.

This also fixes the generator to use `SingleInvoke` for maps. They still
need to be nested one property down to handle secret maps.