Remove ConfigPropertyMap from the engine (#20887)
Remove building and sending `ConfigPropertyMap` as part of run info.
This was an experiment in getting the engine to do type checking and
casting of config so that reading it was easier client side. Turns out
to cause lots of issues because config needs to be dynamically cast at
the read site, for example we don't know if the string "8.0" is the
literal string "8.0" or the number 8, and we can't tell until either
`config.getString` or `config.getNumber` is called in the user program.