Fix import path in instructions when linking a package in Go (#20888)
Reorganize the various `package add` tests, starting with the tests for
providers with a namespace. These now live in
`tests/integration/package-add/package_add_test.go`. Other related tests
will join them soon.

Adding a test for Go for this case revealed a bug in the import
instructions, fix that.