partly leave undangling parent resources to snapshot generation
In `rebuildBaseState`, we try to undangle parent resources.
Particularly relevant for this PR is that we only consider undeleted
resources (resource not marked as `Delete`) as parents. On the surface
this makes sense, however the list of resources we're iterating
through is not always the full list of resources. In particular, for
persisted refreshes in RefreshV2, we only have the old state here.

During the refresh that old state can be marked as "Delete", but we
might also have a "new" state, that's not in the list of resources we
iterate over. This means we incorrectly delete the parent from the
resource in some cases, breaking that relationship.

Since we still have a valid (though potentially less useful) with the
parent being a resource marked as `Delete`, we can just skip removing
the parent at this level, since we simply don't have the full list of
resources here.

When saving the actual snapshot, we later go through the dependencies
again anyway, so we can correct the parent at that level instead,
giving us the correct snapshot.

Note that this always worked correctly in journaling, since we ignore
the results from `rebuildBaseState`, even though there was a potential
bug there when a resource was marked as `Delete`, and there was no
corresponding undeleted resource around, which is also fixed in this
PR.

Note that we can also undo the changes in
https://github.com/pulumi/pulumi/pull/20861, which while they worked,
were incomplete, and this is a more correct solution.