rename copilot to neo
Neo is the new name for copilot, so let's use it consistently
everywhere. Rename internals, command line flags and env variables,
adding backwards compatibility flags/env variables where necessary.

Note that we leave the types in `apitype` as they are, because they
are also used in the service, so renaming them is extra work there as
well, and might not be worth it.