rename copilot to neo (#20898)
Neo is the new name for copilot, so let's use it consistently
everywhere. Rename internals, command line flags and env variables,
adding backwards compatibility flags/env variables where necessary.

Note that we leave the types in `apitype` as they are, because they are
also used in the service, so renaming them is extra work there as well,
and might not be worth it.

Fixes https://github.com/pulumi/pulumi/issues/20873

---------

Co-authored-by: Julien <julien@caffeine.lu>