target: generated skipped same steps for dependencies
When doing a targeted up, we emit same steps for any resources pending
delete, that are not targeted. These same steps are emitted before any
other delete steps. So it is possible for a same step for a child to
be generated before the corresponding step is generated for its
parent.

The snapshot generator works by creating a resource list, starting
with all the resource from any steps, and only then appending all the
existing resources that haven't been removed through any step.

This ordering works as long as steps are created and executed in the
right dependency order, however they are not in this case, as the same
step for the child gets generated before the delete step for the
parent.

These out of order steps mean that the child (or dependent) that was
not targeted and we generated a same step for, now comes before the
its parent/dependency, which is not a valid state for the snapshot to
be in. This state is temporary as we delete the parent later, but the
snapshot generator will error out at the first invalid state, after
writing it, so the users will hit this issue either way.

There's also a second problem here. Since we're deleting the parent of
the resource, that also needs to be reflected, otherwise we still get
a broken snapshot. The same step however doesn't break this
relationship, and thus even if we didn't get the earlier "comes after
it" snapshot integrity issue, we'd still get a "missing parent" issue.

Note that this is an issue only since https://github.com/pulumi/pulumi/pull/20061,
as we did not generate same steps for these resources before at all.

So the proposed fix here is happening in two steps:
- Introduce a new "SkippedSameStep", so we can still emit the same
  step for the display layer, but then avoid recording the step in the
  snapshot manager. The snapshot manager will then be correct, as the
  resource will stay in the old list of resources, and thus come after
  the parent/dependent.
- If there is such a skipped same step, also rebuild the dependencies,
  similar to what we do for refreshes. We need to do that, as
  dependencies/parents of resources can be dropped, but we didn't know
  about that in the step. Thus we need to recalculate the dependencies
  later.