catch panics in the engine, and turn them into errors for fuzzing
The engine ideally shouldn't panic, but it still does sometimes, and
especially during developement. We have a panic recovery mechanism,
but that doesn't work for recovering from panics from goroutines.

This is especially annoying for fuzz testing, as the goroutine panics,
and it's very hard to debug why that happened. Allow catching the
panics in the goroutine and forwarding them to an error channel so the
engine can turn it into an error and return it rather than aborting
prematurely.

We could potentially extend this to catch the errors without the env
variable, which would have 2 main benefits:
1) When this happens to a user, we never get to relinquish the display
   control, so the terminal ends up being messed up until the user
   `reset`s it manually.  This would be fixed.
2) When running multiple tests at once, they all break, and it's
   impossible to tell which one caused the panic. Though we could just
   set the env variable in the TestMain functions, which would have
   the same effect.

For now just introduce this in fuzz testing by default, and then we
can extend it later if we wish to do so.