schema: allow all pep440 constraint for specifying pulumi version in python
We currently allow version constraints like >=version1,<version2 to
specify the required version of the pulumi dependency for python in
the schema. However pep 440 specifies a lot more version specifiers
that can sometimes be useful to use. Implement all of them (other than
`===` which we can't validate).  We still require a lower bound to
always be specified, so we can validate the version numbers. Also add
tests for this.