Add `ReplaceWith` resource option (#20693)
This PR adds a new option - `replace_with` - to the list of options that
can be set on a resource. When resource A sets `replace_with: [B, C]`,
any replace operations on B or C will automatically trigger a replace
option on A. While quite hefty, the majority of this PR is just adapted
code from the `deleted_with` path, and the interesting part is mostly
contained within `step_generator.go`.

Fixes #14072.