Fix closure serialization on Node.js v25.2.0
Node.js v25.2.0 now throws when accessing `localStorage` and the `--localstorage-file` option is not provided. We just happen to be accessing it from the closure serialization code when looping through well known global objects. Since `localStorage` now throws by default, we can simply skip it like we do for a couple other things that similarly throw when you try to access them.