Temporarily pin to Node.js v25.1.0 for tests
Node.js v25.2.1 introduced a breaking change that causes an error to be thrown when the global `localStorage` object is accessed and the `--localstorage-file` option is not provided. Our closure serialization logic hits this because we have some code that loops through well known global objects, and when `localStorage` is accessed we get the new error.

https://github.com/pulumi/pulumi/pull/20913 fixes this, but we have some tests that are still using the currently released npm package.

To unblock merges into pulumi/pulumi, we'll pin CI to use Node.js `"25.1.x"`.