Fix StackReference secrets performance (#20908)
A Pulumi program can use a StackReference to retrieve the stack outputs
of another stack. Internally, the engine will fetch the referenced
stack's snapshot and unmarshal it into a `Snapshot` object. Then the
root stack resource's outputs are extracted from the snapshot and
returned to the referencing program.

As part of unmarshaling the snapshot, the engine will decrypt all
secrets within the snapshot, even though there may not be any secrets in
the root stack resource's outputs.

One limitation is that the referencing stack must be able to decrypt the
secrets of the referenced stack. If both stacks are using the same
secrets manager (such as is the case with the Pulumi service),
everything works fine. If they are not, historically, an error was
thrown, stopping execution of the referencing stack.

https://github.com/pulumi/pulumi/pull/19533 addressed this limitation by
effectively catching decryption errors when referencing another stack,
and inserting dummy values, allowing plaintext values to be shared
freely.

Unfortunately, https://github.com/pulumi/pulumi/pull/19533
unintentionally regressed the performance of secrets decryption such
that batch decryption was no longer used. This results in significant
slowdowns when referencing stacks with many secrets, especially when
using the Pulumi service, as each decrypt operation uses an individual
network request.

This change addresses the regression and improves the overall
performance of StackReference secrets decryption in two ways:

1. The error-catching wrapper used to implement
https://github.com/pulumi/pulumi/pull/19533 now supports batch
decryption, restoring the previous performance characteristics before
that change.

2. Rather than unmarshaling the entire snapshot, and thus needing to
decrypt all secrets in the snapshot, we now only unmarshal the root
stack resource's outputs. If there are secrets in the snapshot, but no
secrets in the stack outputs, there's no decryption to do at all. If
there are secrets in the stack outputs, only those secrets need to be
decrypted, and they will be batch decrypted with the fix from (1).

In the future, we may consider further optimizations, like Pulumi
service changes that pre-decrypt secrets in the snapshot and a new
endpoint for only retrieving a stack's outputs (pre-decrypted).

Fixes #20741