import: actually return errors if we have them (#20904)
Currently in import, despite allowing an error return, we always return
nil. This makes it so that the CLI always exits with a 0 exit code, even
when the import fails, which can be confusing. Return the errors we're
getting, so the user can deal with that information correctly.

Fixes https://github.com/pulumi/pulumi/issues/20901