Avoid cancellations due to do slow processing of requests (#20920)
Follow up to https://github.com/pulumi/pulumi/pull/20917

Update the setting to half an hour.

Also ensure the setting is used for all Python and Node.js GRPC servers.

This option is specific to the C++ gRPC library, and does not apply to
the Go, Java or .NET gRPC libraries.