respect aliases when reading resources with a parent
When resources are being read, we check for whether we have an old
version of this resource by taking its parent into account. However
this doesn't work if the parent is aliased, as we then won't find the
resource in `Olds()`.  This leads to snapshot integrity issues, as the
old resource is left in the snapshot (probably to be deleted later,
but not updated correctly in any case). Fix this by taking the parent
into account correctly.