lifecycletests: log diag events
When there is an error during the lifecycletests it's often useful to
also see the diag events that the engine issued. For example when the
step generator errors, we just get a very generic "step generator
errored" error message, while the diag events actually hold the useful
information.

Logging them makes the lifecycletests a little easier to debug when
there are errors, so do that.