Use RunPlugin instead of policy shim plugins (#20572)
Towards fixing https://github.com/pulumi/pulumi/issues/20402

Passing all config via env vars can run into ARG_MAX limits. Using the
new grpc based RunPlugin approach avoids that, passing config via grpc
instead (which still has limits but are much larger, and don't change
from machine to machine).

To support current NodeJS and Python policy packs we start a proxy
service in the language host so that we can get the stack configuration
via the `StackConfigure` grpc call to the analyser plugin and then use
the information from that to set the legacy environment variables with
which to start the actual policy process up.

This has some extra changes to `Attach` from the first run to watch for the
sub-command exiting before we read a port number off it. In the first iteration
of this change we would hang at that point, now we correctly close off the
stdout pipes and exit the read.