take resources marked as `Delete` into account for targeting
Currently when using the `--target` option for deployment, we don't
take resources marked as `Delete` into account when executing the
deployment.  This can lead to issues when the resource marked as
`Delete` is a child of a resource that is targeted.

Normally in such a situation, we force the user to either specify
`--target-dependents`, or specify all the children as targets
manually, and error out otherwise.  However since we don't take
resources marked as `Delete` into account here, we end up simply
deleting the parent/dependency of a resource, leading to an invalid
snapshot.  See for example https://github.com/pulumi/pulumi/issues/20864

If the child/dependent is marked as `Delete`, we should still take it
into account in the same way for the dependency calculation. This in
turn means we also need to let the user target the resource, otherwise
they can't add it manually to the target list (there is the recourse
of `--target-dependents` of course, but it's better to consistently
allow resources that need to be targeted to be added to the target
list).

This fixes the snapshot integrity issue by removing the possibility of
this operation happening as described in the issue.