use batch encrypter for serializing journal entries
We already use the batch encrypter correctly in the
SnapshotJournaler (the one that turns journal entries into a whole
snapshot and stores them in the persister), but forgot to do so in the
cloud journaler.  This leads to performance problems.  Fix those by
using the batch encrypter here as well.