Add Stash resource (#20819)
This adds a new builtin resource to the pulumi engine, a `Stash`.

A stash simply holds a single value of any type. That value is set when
the stash is created and then never changes. This can be used to record
a value during the first run of a program and keep that saved.

With the `triggerReplacement` resource option, you gain the power to
replace that value on some condition. This can be used to generate, save
and rotate material for the rest of the program. For example generating
ssh keys each month.

This can also be used to coordinate multiple dependencies into a single
hook. For example triggering a replace if either of two VM instances
change, and running a create hook if that happens.

This can also be used to replicate the behaviour of
[`terraform_data`](https://developer.hashicorp.com/terraform/language/resources/terraform-data)
resources for the TF converter. By just using the input field (which is
always updated, same as terraform_data).