use batch encrypter for serializing journal entries (#20991)
We already use the batch encrypter correctly in the SnapshotJournaler
(the one that turns journal entries into a whole snapshot and stores
them in the persister), but forgot to do so in the cloud journaler. This
leads to performance problems. Fix those by using the batch encrypter
here as well.

This fixes the performance problems in
https://github.com/pulumi/pulumi/pull/20990 (I tested locally, don't
want to rebase the branch yet, because there's another issue with
journaling that's being fixed in
https://github.com/pulumi/pulumi-service/pull/34801, and that causes 5xx
errors when we use it)