Fix a panic in delete before replace imports at preview time (#20997)
Fixes https://github.com/pulumi/pulumi/issues/20984

Normally on up we would exit out before we hit this panic because a
replace of an imported resource is already an error, but at preview time
we try to carry on and then we hit this panic because we try to work out
the dependencies of this resource but it's brand new (because we've just
imported it) and so it doesn't exist in the dep graph constructed at the
start of the run. Simple fix of just checking if the resource is in the
dep graph before looking it up. We only do this at preview time so that
in a normal update we still panic if the dep graph is malformed and
missing a resource that ought to be there.