Enhance `up` command help message to include note on template (#20996)
Re slack discussion:
https://pulumi-community.slack.com/archives/C84L4E3N1/p1763573741931959

`pulumi up --help` and
[docs](https://www.pulumi.com/docs/iac/cli/commands/pulumi_up/) mention
template but not how used. Adding minimal description.

```
❯ go run ./cmd/pulumi up --help

Create or update the resources in a stack.
...
Note: An optional template name or URL can be provided to deploy from a template. When used, a temporary project is created, deployed, and then deleted, leaving only the stack state.

Usage:
  pulumi up [template|url] [flags]
```