stop trying to compare journal and snapshot when jorunaling is enabled
Right now, when journaling is enabled, we still try to compare the
snapshot we get from the journaler to the one we get from the full
snapshot manager.  Doing so makes the journaler go slower the larger
the snapshot gets.

Since we haven't found any issues with this comparison mechanism since
a while, we don't need to keep this code that slows us down for the
new implementation.  Remove it to get the best performance.