[sdk/nodejs] Fix serialization errors due to SxS copies of the SDK (#21029)
This change fixes the generated JS proto code to not pollute global
state. We do that by adjusting the find/replace calls to handle the
newer protoc codegen that was introduced in #19588. Without this, when
there are different SxS copies of `@pulumi/pulumi` with different
versions, it can result in serialization errors.

Note: I'd love to add an integration test with SxS SDK versions, but I
couldn't think of a good way to do it without depending on a specific
old version of `@pulumi/pulumi`, which I'm not sure we want to do. I did
manually confirm that replacing the v3.208.0 version of `resource_pb.js`
with the fixed version makes the serialization error go away, with the
repro project mentioned in #21028.

Fixes #21028