Generate .gitignore and .gitattributes files for generated SDKs (#21034)
The commit
https://github.com/pulumi/pulumi/pull/21034/commits/4ba1f8012ecfa491ec6c92e90623a091829ba9cd
has the codegen changes.

Generate a `.gitattributes` file for SDKs that marks the SDK code as
“generated”, and lets for example GitHub collapse changes to these files
by default in the PR view.

Generate a `.gitignore` file for SDKs. This exclude build related files
that should not be comitted.

Both of these changes are meant to help maintaining SDKs, both local and
regular SDKs. The SDKs we maintain often already contain similar files.
For local SDKs, even though you often want to commit the generated SDKs,
it is nice to limit the noise in a PR. We also want to ensure that `git
add ./sdks/` does the right thing and doesn't include `obj` or
`node_modules` or similar, in case these are not gitignored already at
the project level.