Use pytest_asyncio.fixture for test fixtures that require an event loop (#21052)
We still see some flakes for some python tests, depending on how tests
are distributed with pytest-xdist, for example:


https://github.com/pulumi/pulumi/actions/runs/19672452282/job/56345359500?pr=21051

We previously fixed some of this by marking test cases with
`pytest.mark.asyncio`, however test fixtures that use settings and other
pulumi runtime parts also need to have an eventloop. This can be
accomplished with the `pytest_asyncio.fixture` decorator.

Also fixes `test_stack_registers_outputs`, which I think never worked
properly. I verified that the regression this was meant to test does
indeed cause a failure if I revert the fix for it.