Stop passing root and runtime options as args to language plugins (#21054)
Language runtimes can now be used for multiple "programs" (pulumi
programs or plugins) and the correct details for the program in interest
are passed as arguments to the relevant methods, e.g. `Run`, or
`InstallDependencies` both take a `ProgramInfo` struct with the root
directory, program directory and runtime options for that program.

As such it doesn't really make sense to pass these details (root and
options) as arguments when starting the language runtime up, and none of
our language runtimes look at these details anymore.

This cleans up the engine and interfaces to no longer send these
arguments, and thus no longer need to take a `ProgramInfo` as a
parameter to the `LanguageRuntime` constructor.