Fix DIY backend to return plan on update
Update to match the behaviour of the http backend. Without returning the plan, doing a preview during an up won't generate a plan.

When experimental mode is enabled, this would just continue without the plan. We should therefore assert if we're expecting a plan and fail if we don't have one.