Add --strict & --iterate up flags
The current mode of operation of pulumi up runs the program once to approximate the operations which will happen, then, once approved, will execute the program again performing all resource updates required. However, it's possible that the preview was incomplete (due to resources being create in an apply callback), or the program generates a different target state on a different run (likely due to a bug in user-code). This is the purpose of update plans which constrain the actual deployment to a defined set of instructions - the plan.

Update plans are currently utilized by saving a serialized plan during a preview operation, then passing the plan file back in during an update operation. If running an interactive pulumi up in experimental mode (PULUMI_EXPERIMENTAL=true) a plan will be generated in memory and the approved update operation will be limited to the plan. If an unexpected operation is encountered, the deployment will stop.

This change introduces a new update mode via a new --strict flag which triggers the, currently experimental plan behaviour within the preview/up operation.

- Strict mode cannot be used in conjection with --skip-preview as the preview is required to form the plan.
- Strict mode can only be used in non-interactive mode with the --yes option (pre-existing behaviour).

Using the flag --iterate, when an unexpected operation is encountered, in-progress operations will be completed then a new preview will be performed and the user prompted.

- Iterate mode implies running in strict mode as described above.
- Iterate mode can only be used in interactive sessions.
- Iterate cannot be used in conjunction with --yes to avoid getting into infinite loops if there's a re-occurring diff`.