Allow --plan flag in up to generate new plans
Update plans are currently utilized by saving a serialized plan during a preview operation, then passing the plan file back in during an update operation. If running an interactive pulumi up in experimental mode (PULUMI_EXPERIMENTAL=true) a plan will be generated in memory and the approved update operation will be limited to the plan. If an unexpected operation is encountered, the deployment will stop.

This change introduces a new update mode via allowing --plan to be empty which triggers the, currently experimental plan behaviour within the preview/up operation.

- Plan mode cannot be used in conjection with --skip-preview as the preview is required to form the plan.
- Plan mode can only be used in non-interactive mode with the --yes option (pre-existing behaviour).