[sdk/python] Allow older grpcio & protobuf dependencies (#21007)
This change downgrades the generated protobuf code for Python to be
compatible with older versions of the `grpcio` and `protobuf` packages
for users who require older versions of these. Specifically,
protobuf>=3.20.3 and grpcio>=1.68.1.

Two new CI jobs are added to the matrix that run the `test_fast` and
`test_auto` Python tests using the lowest Python dependencies on the
minimum supported Python version (currently 3.10).

I also kicked off a one-off full CI run in
https://github.com/pulumi/pulumi/pull/21015 using Python 3.10 and the
lowest dependencies, and all the tests passed.

Fixes #21067