Remove @types/uuid (#21077)
We were using the `uuid` package in our tests, even though we didn’t
declare the dependency in package.json. It happens to be pulled in by
other packages so it worked.

We did pull in `@types/uuid` as a dev depdendency, but that is
deprecated. The `uuid` library ships its own types.

But, we don’t need to use `uuid` at all, the
builtin[`crypto.randomUUID`](https://nodejs.org/api/crypto.html#cryptorandomuuidoptions)
generates a v4 uuid.