allow setting pulumi backend url for NodeJS automation API tests (#21003)
We currently hardcode the url for NodeJS automation API tests to
`api.pulumi.com`. Internally however it is often useful to be able to
run those tests against a review stack or even a local setup of the
service. Allow setting `PULUMI_BACKEND_URL` to override the default
backend for these tests.