Account for name conflicts when calling resource constructors
When there is a Go resource that ends in `Map` or `Array` and `"generateResourceContainerTypes"` is set, we append `Resource` to the constructor function to avoid conflicts. Previously, programgen was not accounting for this, resulting in bugs like https://github.com/pulumi/pulumi-gcp/issues/2349. This PR fixes the problem by using the same function as programgen to calculate the correct constructor. For example, given the resource `pkg:module:ExampleMap`, we would generate `ExampleMapResource` and `NewExampleMapResource`.

Fixes https://github.com/pulumi/pulumi/issues/17085