journaling: batch entries and send sequentially (#21105)
These changes add support for sending sequential batches of journal
entries to the backend rather than sending all entries concurrently.

Entries are serialized via a buffered channel, and are sent by a worker
that reads batches off of the channel. The maximum time between batches
is controlled by a ticker. Batches are sent when the maximum batch size
is reached or when the ticker fires.

The maximum batch size and the period between sends default to 100
entries and 50 milliseconds, respectively. These values can be tweaked
via the `JOURNALING_BATCH_SIZE` and `JOURNALING_BATCH_PERIOD`
environment variables.

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>