enable journaling in testing (#20990)
As the first step in our rollout plan for journaling, we want to enable
it in our CI tests. This will allow us to get some initial test
coverage. The feature flag for turning this on is already enabled for
the moolumi and pulumi-test orgs. This can be disabled anytime by either
reverting this PR, or disabling the feature flag in LaunchDarkly.

Part of https://github.com/pulumi/home/issues/4404