Assert that conformance tests don't use language features (#21126)
We want to make sure that the conformance tests don't become language
specific, as such we don't allow setting of "Language" options in the
schemas of the providers used here. I've had to point this out on a few
reviews to multiple people, so recording it for ever via this test.