Correctly find the innermost Package/Plugin for `pulumi package add` (#21137)
Previously, `pulumi package add` would fail confusingly when it was run
in a Pulumi
Project that was itself enclosed in a Pulumi Plugin. For a concrete
example, consider this
file tree:

    ./
    ├── examples/
    │   ├── github-reference/
    │   │   ├── Pulumi.yaml
    │   │   └── sdks/
    │   │       └── example-file/
│ │ └── file-0.0.0-x268620d6c62955ff1c7535f520f94fbcaf8d6bfd.yaml
    │   └── simple/
    │       └── Pulumi.yaml
    ├── go.mod
    ├── go.sum
    ├── main.go
    ├── Makefile
    ├── PulumiPlugin.yaml
    └── README.md

When `pulumi package add` is run in `./examples/github-reference/`,
`pulumi` tries to
install a local **go** SDK, when it should try to install a YAML sdk
file. `pulumi package
add` would flatly prioritize the enclosing `PulumiPlugin.yaml` over the
`Pulumi.yaml`,
even though the `Pulumi.yaml` is closer to where the command was run.

This commit changes the behavior of `pulumi package add` to look for
both
`PulumiPlugin.yaml` and `Pulumi.yaml`, taking the innermost one if both
are found.

Fixes https://github.com/pulumi/pulumi/issues/21136