[sdk-gen/go] Fix generation of lifted single-value calls in parameterized SDKs (#21115)
Fixes #20744
Fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/3247

### Description

In the issues above, we are dealing with a SDK-gen problem in Go,
specifically for a generated method/call that is parameterized and the
schema has option `liftSingleValueMethodReturns` enabled.

### EDIT: testing via a SDK-gen schema rather than conformance tests 

The following is no longer relevant since conformance tests removed

```
The fix is actually pretty straightforward in Go SDK-gen but to properly conformance test it, this PR:
 - extends the `call` provider with `map[string]any` return type method called `configData` and tests it in `l2-provider-call-explicit` (simulating the type of the problematic `terraformConfig()` method in reported issues)
 - adds `liftedcall` provider and associated `l2-provider-call-explicit-lifted` which is exactly the same as the `call` provider but with a schema that enables `liftSingleValueMethodReturns=true`
 - `liftedcall` is implemented as a parameterized provider but doesn't do anything with the parameterization i.e. value is irrelevant but generated SDK is a parameterized one

Note for Python and NodeJS: this PR excludes `l2-provider-call-explicit-lifted` from the conformance testing because program-gen doesn't handle `liftSingleValueMethodReturns` so it doesn't simplify `method().result` into `method()` and the program fails. It can be handled in a follow-up PR via program overrides similar to what Go language tests do now but opted not to make the PR any bigger
```