Enhance the replacement trigger test with unknowns (#21139)
The replacement trigger option can be passed `unknowns` as well as
normal values, these show up at preview time as replacements. This
enhances the conformance test to check that an unknown value can be
passed in and shows a replace at preview time.

Would be good to also extend this test to ensure secrets work correctly,
but I'll do that after this is merged.