Don't install dependencies for nested git based providers on failure (#21148)
This code was originally added in
https://github.com/pulumi/pulumi/pull/18322. As of
https://github.com/pulumi/pulumi/pull/21119, we no longer allow multiple
components to share a git repo. That means we will never encounter the
situation described in the comment:

The plugin is already installed. But since it is in a subdirectory, it
could be that
we previously installed a plugin in a different subdirectory of the same
repository.

We can now safely remove this check, making git based providers more
similar to other providers.

Stacked on https://github.com/pulumi/pulumi/pull/21119