only wait for update check if we prompt (#21120)
We always want to send the update check request to the service, so we
can cache the results locally. However this can slow some very fast
commands down, without the user knowing why. We don't really need to
wait for this for fast commands, if we know we won't prompt the user.

This brings the time for `pulumi version` from:

Time (mean ± σ): 481.7 ms ± 105.3 ms [User: 177.9 ms, System: 29.9 ms]

to:

Time (mean ± σ): 94.9 ms ± 10.8 ms [User: 123.6 ms, System: 14.9 ms]

Note that we'll still send the request to the server, we just won't wait
for the response, and the command will still be slower if we actually
want to show the prompt.

This was prompted by the recent ideation conversation about how some
commands that should be really fast are still sometimes slow. This
should still allow us to send the request to the service, without the
downside of having to wait for the result.