update step generation documentation (#21153)
Try to make the step generation documentation a bit more accessible,
shortening some sections, and adding more mermaid diagrams to explain
it. It felt a bit intransparent to me the first time I read through, so
hopefully this is an improvement, but also happy to hear otherwise.

(Note that the diagrams are generated with AI help, though I reviewed
them to the best of my knowledge)

Fixes https://github.com/pulumi/pulumi/issues/20981