startup perf: get user info and default org async (#21176)
During the startup of pulumi programs, we currently get the user info
(only if not cached locally), and the default orgs from the cloud. This
is happening synchronously at some point during the startup process.

We can improve performance here by requesting these as soon as the
backend is initialized in a promise, and then just getting the result.

Locally this speeds the operation up by around 500-600ms (this is just
up-ing a stack with a single resource, that doesn't change).

Before:

```
$ hyperfine -- 'pulumi up --skip-preview -s another'
Benchmark 1: pulumi up --skip-preview -s another
  Time (mean ± σ):      6.616 s ±  0.853 s    [User: 1.730 s, System: 0.256 s]
  Range (min … max):    5.895 s …  8.720 s    10 runs
```

After:

```
$ hyperfine -- 'pulumi up --skip-preview -s another'
Benchmark 1: pulumi up --skip-preview -s another
  Time (mean ± σ):      6.065 s ±  0.500 s    [User: 1.695 s, System: 0.262 s]
  Range (min … max):    5.456 s …  7.050 s    10 runs
```

This is something we can do for
https://github.com/pulumi/pulumi/issues/3671, without any service
involvement yet.